added httpmock server library to service tests. first network service integration test passes.
