+++ test/service/server/httpserver.cpp
@@ -1,23 +1,7 @@
 #include "httpserver.h"
-#include <QtTest/QtTest>
-#include <QtTest/QSignalSpy>
-#include "requestmanager.h"
-
-/// Port number server is tried to listen on
-/// Number is being incremented while free port has not been found.
-static const int server_port = 8080;
-
-/// Server started in the main().
-static httpmock::TestEnvironment<httpmock::MockServerHolder>* mock_server_env = nullptr;
-
-static QString getServeUrl() {
-    assert(nullptr != mock_server_env);
-    const int port = mock_server_env->getMock()->getPort();
-    QString url;
-    //std::ostringstream url;
-    url.append("http://localhost:").append(port).append("/");
-    return url;
-}
+//#include <QtTest/QtTest>
+//#include <QtTest/QSignalSpy>
+//#include "requestmanager.h"
 
 HttpServer::HttpServer(int port) : MockServer(port) {
 }
@@ -53,22 +37,22 @@ HttpServer::processHeaderOutTest() {
 }
 
 TEST(MyTest, dummyTest) {
-    auto parent = new QObject;
-    QNetworkReply *reply = nullptr;
-    auto bsfRequestManager = new RequestManager(parent);
-
-    QSignalSpy spy (bsfRequestManager, SIGNAL(httpCallReady()));
-    QString url = getServeUrl().append("test");
-
-    bsfRequestManager->sendRequest(url, reply);
-
-    ASSERT_EQ(spy.count(), 0);
-
-    QVERIFY(spy.wait(1000));
-
-    ASSERT_EQ(spy.count(), 1); // make sure the signal was emitted exactly one time
-
-    ASSERT_EQ(reply->readAll(), "payload test");
+//    auto parent = new QObject;
+//    QNetworkReply *reply = nullptr;
+//    auto bsfRequestManager = new RequestManager(parent);
+//
+//    QSignalSpy spy (bsfRequestManager, SIGNAL(httpCallReady()));
+//    QString url = getServeUrl().append("test");
+//
+//    bsfRequestManager->sendRequest(url, reply);
+//
+//    ASSERT_EQ(spy.count(), 0);
+//
+//    QVERIFY(spy.wait(1000));
+//
+//    ASSERT_EQ(spy.count(), 1); // make sure the signal was emitted exactly one time
+//
+//    ASSERT_EQ(reply->readAll(), "payload test");
 }
 
 TEST(Server, launchFailure) {
@@ -83,14 +67,12 @@ TEST(Server, launchFailure) {
 
 
 int main(int argc, char *argv[]) {
-    QApplication app(argc, argv);
+    //QApplication app(argc, argv);
     ::testing::InitGoogleTest(&argc, argv);
     // startup the server for the tests
     ::testing::Environment * const env = ::testing::AddGlobalTestEnvironment(
             httpmock::createMockServerEnvironment<HttpServer>(server_port));
     // set global env pointer
-    mock_server_env
-            = dynamic_cast<httpmock::TestEnvironment<httpmock::MockServerHolder> *>(
-            env);
+    mock_server_env = dynamic_cast<httpmock::TestEnvironment<httpmock::MockServerHolder> *>(env);
     return RUN_ALL_TESTS();
 }
